Add button to Reset Bulk Update values
The form on the bulk update page was keeping its values after updates,
in case you need to run it multiple times. However, the form is large,
and if you are doing several different bulk updates, it was possible to
accidentally apply a previous change to a new bulk update.

This commit adds a "Reset" button to the bulk update page that will clear
all the values in the form, allowing the user to start fresh.