Remove simple-slack-notify GitHub action
This action was used to send notifications to Slack when a build/test
finished. It was deprecated and there is no new version compatible with
node 20, which is required by the new GitHub Actions.

We have the alternative of using the GitHub App from Slack, which also
provides notifications for workflows, so we can remove this action.