Warn if queue addresses are the same as a user address
Some admins may want to set the queue's comment or correspondence
address to the same as a user's email address. Such settings can cause
email loops and complications with email signing and key management.

This commit adds a warning to the admin interface when the queue's
comment or correspondence address is the same as a enabled user's email
address.