Always pass import flag within LDAP import
When constructing a user object in the LDAP import, the import flag was
not passed through as it is elsewhere in the import process. If this
flag is needed in the LDAP field mapping, there was no guaranteed
availability.

Explicitly passing the import argument to _build_user_object ensures
that LDAP mappings will always be aware of whether this is a live or
test import.