Merge existing mappings on update when Name is absent
The Name argument was added in 3e9651f2c1, so admins can delete some items
on Advanced page. The code has a bug that it accidentally deletes mappings
of other lifecycles on Mappings page. This commit fixes it.

Technically, we shouldn't touch existing mapping items before merge when
Name is absent.