Allow to get the Stage of a Scrip passing the queue id as argument
Previously, the Stage method of Scrip only allowed to get the Stage of
the scrip by passing a ticket object as argument.

This commit updates the Stage method of Scrip, allowing it to receive
queue_id named argument alternatively to get the Stage of the scrip
informing the queue directly and not only the ticket.

This will be useful to get the Stage of a Scrip directly for new
features like the full Scrip copy feature.