Introduce scrip copy functionality
Previously, to copy the parameters from an existing scrip to a new one,
you have to manually recreate each field.

This commit adds a "Copy" link to the Modify page for a scrip.  By
clicking the link, RT will clone the scrip and disable the new version.
It will also copy the applied objects and stages.