Use bootstrap's default fonts
Bootstrap puts effort into using the best fonts available on each
device, so leverage their work and use the default bootstrap
fonts.

Also use bootstrap to set the default base font size. The size
change from 13px to the typical browser default of 16px is a
big change, so we use 0.9rem to not be quite so big.