Align right justified elements with bootstrap card edge
Previously right justified items in the titlebox were
positioned to the right of the bootstrap card right
edge, so it seemed to hang over. Align it to make it look
more consistent. You can see this with the version number
on the login page and the top gearbox on ticket display.

Also remove the redundant 15px padding on cards since
bootstrap already sets the same value. Keep the override
of the top padding.