If the submitted Due value is unchanged skip the message
If a submitted value is the same as the current value for
Due on the reminder, we skip the update. Showing the user
a message that the value is invalid is misleading.

The updated Reminders format includes existing Due values
and we don't want to show invalid messages for each.

We could update the message to be more accurate and say
"that is the existing value", but that's not really helpful
if the user was updating another field. Instead, remove the
invalid message and do nothing if the new value is the same
as the current Due.