Allow a little more delay on menus
I am sometimes annoyed by accidentally opening a menu by
mousing over it, then having to wait for it to close before
I can access the page under the menu. However, the other use
case is opening a submenu and missing by just a bit and
having the entire menu close up.

The previous delay setting of 100 was too fast for the second
case. As soon as your cursor moved off the menu, it was gone.
Setting a little higher gives just enough time to mouse off
a submenu and back on before it closes up.