Allow user override in RightsInspector Search method
The Search method in RT::RightsInspector was originally written to
support the AJAX call from the Rights Inspector admin tool, so it was
always run in the context of the logged-in web user, accessed via
$self->CurrentUser.

That method is useful for performing rights searches from code as well,
where CurrentUser may not be defined or where you might want to run the
search as a different user, like the SystemUser. Allow a user object to
be optionally passed in to handle this use case.

Use the lowercase "user" to be consistent with the other existing
arguments for the method.