Resize ticket timer to remove scrollbar
Several small changes to the rendering of the ticket timer
to prevent the scrollbar from appearing. The current height
removes the scroll bar in all browsers tested except for
Opera. The height needed to prevent the scrollbar on Opera
made the other browsers look worse with a big blank area at
the bottom.