Dynamically resize ticket timer popup for zoomed in screens
The width/height args passed to window.open do not auto scale for zoomed
screens, this may cause the pop up window to be smaller on zoomed-in
screens.

window.devicePixelRatio might contain zoom level but sadly not it's not
always true. E.g. it doesn't change on Safari no matter what zoom level is.

On the other hand, element's clientWidth is reliable and we can calculate
the zoom from it and adjust window size accordingly.