Update queue cache tests to load pages with a queue dropdown
These tests were originally written to use the queue dropdown
that used to be in the main RT header next to the "Create ticket in"
button. They continued working after we moved that to the ticket
create page because the Quick Create portlet also has a queue
dropdown. When that was converted to htmx, the Mechanize tests
then failed.

Update these tests to navigate to the ticket create page where
the queue dropdown now lives.