Fix query builder to handle unselected custom fields
Query builder failed when a value was entered for a custom field without
selecting the custom field name.

This caused the query to generate invalid statements like
'ARRAY(0x55e289486b98)' LIKE 'value'.

This commit makes the query builder ignore the custom field value if the
custom field name is not selected.