Support base and target validation with a link type
The circular link validation in _AddLink previously assumed
that all base and target parameters would be ticket ids.
Values can also be passed in the form 't:1' and that would
cause warnings and not perform the validation.

Use the URI methods to get an id from any valid base or
target values.

Fixes: I#31399