Dismiss modals on submit
This wasn't necessary before htmx changes. Since we do not reload the whole
page, we need to dismiss modals on submit, otherwise the main body would be
greyed and not focusable.

Note that previously we closed saved search options modal only when there was a
successful change, which isn't enough. With this commit, the old behavior wasn't
needed any more.