Set the default 'unset' language hint to the browser default
The previous code called LanguageHandle, which would return
the current lanauge if it was set. So if your browser was US English
and you selected French and saved, the unset option would show "- (French)"
because that was set. But this is incorrect, because if you selected
that option and saved, the language would be US English.

Set the hint on the unset option directly from the value passed in
the Accept-Language header instead.