Add option to show user text as avatar tooltip
In some places where user is displayed, presenting the avatar and the
user text can be redundant and take up too much space.

This change adds an option to show the user text as a tooltip on the
avatar. This way, the user text is still available, but it doesn't take
up as much space.