Ignore negative answers of equivalent object cache
This is mostly a typo, negative short answers should be skipped.

Previously if a principal got rights applied to an object but not globally,
running the following code returned the wrong result, e.g.

    $user->HasRight( Object => RT->System, Right => 'CreateTicket' );
    $user->HasRight( Object => $queue,     Right => 'CreateTicket' );

Even if the user had "CreateTicket" applied to $queue, the second HasRight
call above returned false.