Remove Basics tab on tickets because inline edit allows updates
The Basics tab was original the editting interface for all of
the ticket core fields and custom fields shown in view modd
on the ticket Display page. The Display page now has inline
edit which provides forms covering all of the fields on the
Basics page, so the stand-alone page is no longer needed.