Update tests for new Time widget and removed Basics
For t/selenium/basic.t, the tests for the Basics page
are removed because the Basics page has been removed.
Other tests were migrated to use Jumbo, but selenium shows
spurious errors submitting changes on the Jumbo page, so
the tests are just dropped. Similar tests are run on Jumbo
in t/web and t/selenium/ticket_inline_edit.t covers the new
primary way to modify tickets, so we still have coverage.