Show all available dashboards for superusers
Initially in f6cb772b25(in 2011), we excluded dashboards of groups which
current superuser does not have explicit "SeeGroupDashboard" granted on from
menu.

As users can customize dashboards in menu now, the hard limit is not
necessary any more. This change also makes the behaviors of saved searches
and dashboards more consistent.