Update MaxAttachmentSize documentation
Both items of missing information (the interaction with the dropzone
widget and the internal reduction when using some databases) have
surprised at least one RT admin.

Without these being documented the only way for an RT admin to discover
them is by reading the code, which they shouldn't be required to do.