Don't return a disabled Default Queue
If a user has a Default Queue selected in preferences and that
queue is disabled, the Create Ticket page showed an error
and did not allow the user to select a different queue, blocking
them from creating new tickets. Detect if any of the Default
Queue options is disabled and return undef, allowing the Create
Ticket page to load and present other queues for the user to select.