Use LabeledValue for custom field edits
As part of this, move the LabeledValue call into EditCustomField.
The new layout accepts FormText to show a current value for dates,
and much of the calculation of the correct default value was
already in EditCustomField. This refactor ended up removing some
duplicated code as well.

Also, default Column to undef to allow width to fill the surrounding
container by default.