Update user preferences page to use LabeledValue
The form elements on the preferences page are rendered using
the shared Widget templates which are used in other places
also. Add LabeledValue as a new option for display to preserve
the previous layout in other areas.

The new format compresses the options by displaying the label
on top of the option, so also convert to a two column
layout.