Add loading lazy attribute to img tags in transactions
A ticket with a long history can have dozens of img tags,
and if the full history loads on ticket display, the RT
server is hit with dozens of requests because each is an
attachment that needs to be served by RT.

Add the loading="lazy" attribute to img tags to make the
browser wait until the image is in the view port before
loading it. This newish attribute is now supported in all
major browsers.