Use GMT in Time::ParseDate to adjust for user timezone
The tests would only pass if the server running the tests was in the
America/New_York timezone.

The issue was with the Time::ParseDate::parsedate method that was determining
the timezone from the server environment.

The fix is to call Time::ParseDate::parsedate with the GMT argument and manually
adjust for the user's timezone offset.