Reformat ticket basics to improve layout
The new label/value display creates more horizontal space,
so reformat the information in Basics to extend into some
of that space. The layout is still responsive as the screen
gets smaller for tablets or mobile.

Also update edit pages to display inputs in the same order
as the display. Add all basics options to the update page
to make it convenient to update priority and SLA.

Restore Owner to the display page for consistency and to make
it easy to see and edit Owner.