Avoid compilation errors if GnuPG::Interface is not installed
Although we hide the page at Admin/Tools/GnuPG.html when GnuPG is
not active in the RT configuration, it is possible to go to the
page directly. Previously this would result in a complilation
error when RT couldn't load GnuPG::Interface. The test
t/web/compilation_errors.t would also fail, even when running
tests with GnuPG disabled.

Check for GnuPG::Interface and provide a standard page with
a message if it is not installed.