Restore Stacked selection on Charts
In 08cc6792bd, we removed the fieldset tags to standardize
the group by selections. Some JS on the page triggered based
on the fieldset tag, so selecting Stacked charts stopped working
after that change.

Add a custom class to each group by row and update the JS to
restore the Stacked selection.