Revert to the default bootstrap border radius
In RT 5 we had a custom border radius from the first iteration
of the elevator theme. Revert to the default bootstrap
layout as a change for RT 6 that is more rounded and
looks more "modern".

The additional color changes are from updates to the sass
build utility. They shouldn't change the actual colors in
RT. Details can be found in their change log.

    https://github.com/sass/dart-sass/blob/main/CHANGELOG.md#1790