Handle htmx 400 and 500 server errors with custom response-targets
The response targets extension detects 40x and 50x server
errors and puts the error message response into the location
you define, possibly a designated error message div on the
page.

In our case, for GETs, we instead want to replace the server
error message with a custom formatted message, and put it
in the location that was trying to fetch content. We do
this with a small modification to the default extension
behavior.