Display a message on htmx network errors
With a traditional web page, the browser displays an error
if it can't reach the requested server because of network
errors. With htmx, the request is in JS, so the browser
doesn't know to show an error.

Detect these network errors and display a growl message so
the user knows why their click or other action isn't
working.