Allow variable interpolation in SQL heredocs
Commit 3e3b174f6e added a variable that holds the Groups table name so it can be
quoted correctly for MySQL 8.

The SQL heredocs used single quotes though, which means the variable was not
interpolated but instead taken literally. This caused an SQL error as the table
does not exist.

Changed the SQL heredocs that use the variable to use double quotes so the
variable is interpolated correctly.