Ensure a transaction exists for adding CGM records
In commit 304fd96a81 there was a change to add CGM records all at once at the
end of the import process, instead of adding as other records were created.

When using no-auto-commit this can create an error as a new transaction is not
started and the CGM records are not committed.

Using no-auto-commit also causes Commit to be called when there is not an active
transaction.

Added logic to only call Commit if there is an active transaction and to start a
transaction when adding CGM records, if there is not already an active
transaction.