In transaction history, show avatars only for discussion
Showing the user avatar on every transaction makes the history
very busy, and it doesn't look good when there is a long series
of ticket metadata updates that are a single line.

Show only on transactions that are related to conversation,
where it can look like that person is talking. This also
helps these transactions stand out in the history.