Add aria-label attribute to GetSVGImage when Title is sent
Each column header on the ticket results page has a filter icon.
Although this icon has a title, it was not being read correctly
by screen readers like NVDA, which would speak "void zero."

This change adds an aria-label to any generated SVG when a
Title is provided. This improves screen reader compatibility
and ensures the icons are read meaningfully.