Improve layout of theme page elements
Move the Customize elements up into the same column
as the logo upload to fill the large empty white space.

Add a border around the logo and an hr to give some
clearer defintion to the elements.

Also remove the custom theme css that is no longer needed
because of bootstrap defaults.