Stay on current form and show invalid fields on error
Since the validation call with htmx is run in the background,
we return validation errors with a 422 status code. This is
detected by events in htmx and the UI is updated to indicate
which fields had errors.

Using 422 for this case isn't accepted by 100% of htmx
users, but it seems most are using this pattern. It also matches
what Rails is doing, and they use turbo, which is similar
to htmx.