Add new callbacks to modify the main page navigation menu
We no longer build the main page navigation on every request
in the same mason template as the page menu, so extensions
that used /Elements/Tabs Privileged and SelfService to modify
the top page menu will no longer work.

Add a new callback that can be used for this purpose.