Load images in history by SHA if possible to enable caching
Tickets can often accumulate many copies of the same image
from company logos in email signatures. Previously RT loaded
each of these as a unique image, resulting in many requests
to the RT server for what is actually the same image.

Load the image based on the SHA instead. This will result in
the same image URL for all equal images, even if we have stored
multiple copies. The browser will then cache the first image
and re-use it, removing all of the additional requests.

The image will also be cached on other pages, saving more requests.