Filter mason components requested from views
We have a few dhandler components in /Views/, which generally are a wrapper
of components in /Elements, /Ticket/Elements, etc.

Previously we allowed any components to be called there, which is a bit
risky considering some components might contain sensitive info. This commit
implements a way to identify whether the component can be called or not.

Homepage components are allowed by default. In addition, only components
with attribute "directly_accessible => 1" are allowed.