Switch to POST for widgets on create/update pages
For create/update pages, forms could be re-rendered on submit, instead of
creating/updating tickets/assets, e.g. when user selects an article on
ticket update page.

When this happens, it's not appropriate to use "GET" to render each widget
with all posted data, which might be too big. This commit switches to "POST"
accordingly.