Stop using the session to pass report info to JS Chart
When using GD to create charts, we created an img tag that
would be loaded when the chart was rendered. To avoid rebuilding
the report, we stashed data in the user session that could be
accessed when the browser made a request from the img tag.

JS charts are called directly in mason and don't use an img
tag, so we don't need to use the session to pass the data.
Pass the columns and report directly as part of the mason
call.

This avoids issues we have seen where some kind of contention
with the session could cause charts_cache entries to not be
deleted. The session could then grow in size until the user
logs out.