Let Scrips handle different object types
The application of scrips to objects was already handled in a pretty
generic way via ObjectScrips, so most of what we need to do here is give
RT::Scrip the LookupType role, and make sure to use classes derived from
the LookupType, instead of assuming tickets and queues.

For this the LookupType column is added to the Scrips table, and
existing records take the RT::Queue-RT::Ticket LookupType.

Scrips on objects other than queues are forced to use only global
templates, because we still need schema and other changes to let them
have override templates.