Retain extra widget configuration after a : in widget name
Specifically for CustomFieldCustomGroupings, retain the grouping
that might be in the widget name after a :, like
"CustomFieldCustomGroupings:Custom".

Previously the "selected" grouping did not show up in the edit
modal because the grouping part of the widget name was being
cleared.

Also create a new file for page layout functions since it seems
likely we will have more.