Bypass ACL for Scrip->LookupType
In order to see if user has the ShowScrip permission, you need to query
ObjectScrips for items in our CollectionClass, and that depends on the
LookupType. t/security/CVE-2011-2084-modifyscrips-templates.t contains
an example of ending up in deep recursion without this.

Looking at how other classes get around this problem:
- CustomRole just doesn't check permission in the _Value method.
- CustomField has the "ContextObject" scheme. I suppose this could be
  done for Scrips if desired, but it might be overkill.
- More of a system-wide solution would be to bypass the ACL from within
  LookupType.pm, like was done for Lifecycle in
  e59c1f2c0f4a412b8d96142ad54a20bc3ccd1887.