Fix page menu links for Scrips
Enables the "From" argument for non-queue scrips, and automatically
links back to the correct subdir for the object we started from.

Added helper function _AdminPathFromLookupType function, which just
pluralizes the RecordClass, e.g. Queues. Unless it's from Articles or
Assets, in which case it also needs to go under that subdir.