Unset hx-vals right after widget's htmx load request finishes
We only need it in initial load request and no more.

Note that hx-vals is an inheritable attribute and all children's htmx
requests inherit values from it, which is annoying in most cases. E.g. if
the widget contains a link that points to another ticket, we would have to
explicitly set hx-vals="unset" to the link to prevent it from inheriting
parent's hx-vals.