Show empty input for easy search on tom-select dropdowns
One nice feature of tom-select is it provides a search
feature for dropdowns so you can start typing to find the
item you are looking for. However, by default when a dropdown
has an existing value, tom-select shows the value in the input
box and puts the cursor to the right. You can start typing to
find a new value, but it's confusing to see the old value
still displayed as you type to search for a new value.

Clear the value on focus so you can start typing immediately.
If no value is selected, restore the original value. This makes
dropdowns behave the same as RT 5.