Submit inline date changes selected from date picker
When dates are added from the date picker, the default
change event doesn't fire because there is no focus to
blur change. Watch changes from the date picker and
mark the field as changed on update.

RT 5 had a similar event on the date picker and this
reproduces that functionality.