Remove submit on close from RT 5 datepicker
In RT 5, the date picker had a Done button and date
changes were automatically submitted when done was
clicked. The new date picker has an X rather than
Done, so it's not clear that clicking that will not
only close the picker, but submit the value. Users may
want to check and edit the value before submitting, so the
auto-submit may be surprising. Remove the non-functioning
auto-submit code for now rather than updating it to trigger
on the X click.

Users can still select a value and click the green checkmark
directly.