Canonicalize widgets for both dashboards and tickets/assets layout edit pages
For dashboards, widgets are saved as hash references that contain essential
items only, we need to expand them to get more meta info like label,
section, etc.

For tickets/assets, widgets are saved as strings that are composed by widget
name and an optional argument like "CustomFieldCustomGroups:Default", we
need the saved version to get argument.