Make sure saved widget collapse states are saved correctly
In RT 6, widgets are individually loaded. This brings an edge case that
sometimes the widget that handles the collapse events isn't the widget the
collapse state is bound to.

E.g. on ticket display page, "Basics" and "Custom Fields" widgets are in
"Ticket metadata" widget. Assuming "Basics" widget got collapsed, then on
page load, when RT tries to automatically collapse it(it happens on each
htmx load), the "Basics" widget might not be fully initialized yet, in which
case the event handler is actually its ancestor widget "Ticket metadata",
and it causes RT to wrongly save "Ticket metadata" as "collapsed" too.

To prevent it from happening, here we use evt.target.id to get the element
the event is initially dispatched to, instead of the one that actually
handles the event.

This commit also stops event propagation and load collapse states for new
elements only, to reduce unnecessary work.