Restore inline edit as the default for the ticket Owner format
In 15955a09cf, we removed inline edit for the default Owner format
for tickets because it was expensive to build the Owner dropdown
for all of the tickets in results and this made the results page
slow. We now have a delay mode that loads the Owner input only
when the user clicks on it, so there is no penalty when the results
load. Restore inline edit as the default with this Delay mode
active.