Align tomselect dropdowns in search filter according to modal position
By default, tomselect dropdowns are left aligned, when options in dropdown
are too long to fit in the modal, they overflow automatically to the right.
The issue is when modal itself is right aligned, then the overflowed content
is actually invisible.

This commit changes tomselect dropdowns to be right aligned when modals
appear at the right side of screen, so content can overflow to the left.