Do not override passed in $Transactions in /Elements/ShowHistoryPage
Previously we overrode passed in $Transactions to the given object's all
transactions, when the URL contained the ReverseTxns param. This caused RT to
repeatedly render the ticket's full history on scroll.

ReverseTxns is only used to get given object's full and ordered
transactions. There is no need to order passed in $Transactions, which might
only contain partial transactions, and is supposed to be ordered already.