Add selectpicker class to display columns select
The display columns select was being initialize directly by adding a call to
initializeSelectElement in some JavaScript code. This was causing an issue when
the query builder was reloaded by htmx and the underlying tom select control
would become visible when it should be hidden.

By adding the selectpicker class the display columns select correctly gets
initialized even when the query builder is reloaded by htmx.