Warn unknown domain on web UI
For CSRF protection, we check Referrer header and forbid update requests
from unknown domains. For ajax/htmx requests like inline edit, it's
complicated and inconvenient to show another page or modal and ask users to
confirm the operation. Here we show a warning message on normal pages to
instruct people to go to the correct domain instead.