Fix Shredder Attachments plugin validation test
Some files could not be found while using the Attachments plugin in
the Shredder due to an incorrect validation regex which was not
considering valid characters such as dash (-) and other valid special
characters in filenames.

This commit updates the regex to ensure it correctly matches all valid
characters in a filename.