Fix duplicated submissions for ticket description inline edit modal
Previously the description inline edit form on search results pages was
submitted twice on save, from submitInlineEdit and htmx's default behavior
(the form is created with hx-post), respectively.

Even worse, as submitInlineEdit was called on the form's "submit" event, and
calling submitInlineEdit triggered another "submit" event, this caused an
endless recursion.

This commit adjusts the "Save" button to be a normal one, and registers
submitInlineEdit to the button's "click" event (quite like other inline edit
forms). In this way, the form submits only once now.