Exclude unnecessary blessed arguments in hx-vals of widgets
Arguments passed to ShowWidgets are not supposed to contain Perl objects.
But there are cases that we might pass arguments indiscriminately, of which
code is quite straightforward: ARGSRef => \%ARGS.

An example is ticket update page, where we pass $TicketObj to Display.html.
This would cause a JSON encoding error when submitting an unchanged form on
the ticket update page.

This commit filters out blessed objects to make JSON encoding happy.