Access queue name directly when checking configuration settings
When accessing queue name to check for configuration settings,
the check will fail if the current user doesn't have SeeQueue
because the name will be undef. This can happen if a user is
granted ShowTicket directly without SeeQueue, for example.

In these cases, the queue name is not being revealed to the
current user, so access the value directly to allow the
configuration check to run correctly.