Improve dropdown interactions with the dropdown_input plugin
This tom-select plugin creates an input box as the first item
in the list when a dropdown is selected. Users can then type
in the box to search the list of dropdown items.

Enabling this also changes the behavior when an item is
selected from the list. Instead of putting a cursor next to
the selected value, the new value is set and there is no
cursor, more like a typical dropdown. This seems more natural
and more like what users would expect from standard dropdowns.