Remove tickets Edit menu
0be5927 introduced a new Edit menu for tickets, moving menus that were
previously at the top level of the page menu to save some space on the
page, as most of the information that could be edited on those submenus
could also be edited in the ticket display page with inline editing.

Most of the previously moved submenus (such as Basics, Links, and
Dates) were recently removed, leaving the Edit menu nearly empty.

This commit removes the Edit menu from the ticket page, and moves
the remaining items back to the top level of the Page menu.