Enable htmx for ticket and asset search filtering
Remove the separate filter form and update the forms for
each search filter to build dynamically when clicked.
htmx can now handle the submit on each form.

This fixes some issues with the modal backdrop not
clearing after submitting a filter term with the Enter
key. It also fixes cases where a search would be lost
and the default id > 0 search would be triggered.