Tweak the position of tomselected inline edit inputs on search result page
Tomselected inputs' width may exceed the column's width that thus overflow
off the screen when the column is the rightmost one. This commit pushes them
to the left a bit to get rid of the overflow.

Just like in filter modals, here we also align dropdown options to the right
to give them enough space to show up.