Use high resolution time for Queue/Catalog/CustomRole/Lifecycle cache flags
When we update cache and a related object like a queue changes at the same
second, previously we had to update cache again as we
couldn't tell if the current cache contained the latest changes. By
using high resolution time, this should happen rarely.

This fixes a test failure happening now and then in t/web/session.t that is
caused by the above edge case.

Note that ConfigCacheNeedsUpdate is not changed intentionally because the
time comparing it to is Configuration's LastUpdated column, which isn't high
resolution.