Merge extension config meta with existing meta
If extension config values are loaded from the database when RT starts
RT::Config::META will already have entries (like the Source entry) for the
config options when the extension calls RegisterPluginConfig. The values
passed to RegisterPluginConfig should be merged instead of overwriting the
existing entries.

This fixes the issue that extension configs set from the database were shown
as "core config" on system configuration page.