Split multiple tomselect initial value on delimiter
When setting the initial values for a multiple value tomselect control the
values need to be split on the delimiter to avoid adding a value that includes
all delimited values.

Since ", " is the delimiter for multiple value inputs, we should process
default values in EditCustomFieldAutocomplete in the same way. This fixes
autocomplete's default values when creating a related ticket from Linked
widget.