Default to the current class for existing articles
When showing the edit page for an existing article, the current
article class should be selected in the class dropdown,
not the system default for a new article.

In the strange case that Edit.html receives both an id and
a Class, we use the class as the default. We assume the intent
would be to load the article from the id, then set Class to the
value provided in the Class argument.