Convert CKEditor toolbar configuration to extended format
The CKEditor toolbar configuration has a basic format and an
extended format. The extended format with toolbar entries defined
in "items" is required if you want to use other toolbar options
like shouldNotGroupWhenFull.

This commit does not add any extra options, but defaulting to the
extended format makes it easier for users and extension authors
to add options in the future.