Bypass ACL cache for owner validation on ticket queue change
On ticket display page, we pre-cache ACL to ensure current user can complete
all changes in sequence, preventing failures due to permission revocation
from earlier modifications.

When verifying whether the ticket owner shall retain the ownership in the
new queue, we need to ignore ACL cache to ensure an up to date result. This
fixes an issue that when current user is the ticket owner, changing queue
from inline edit on ticket display page retained the owner, even if the user
didn't have the permission in the new queue.