Log SavedSearch calls for missing ContentObject records
SetContent calls ContentObj to determine if content already
exists, so missing records were not logged because it was
normal for this case. However, this also hid possible issues
in the DB if a ContentObject record was missing for a SavedSearch.

Update the public method to log any failure to find or return
content. Keep an internal method for SetContent.