Exclude id from NOT NULL checks for user autocomplete
In d9e88747c3 we updated autocomplete interfaces to use
tom-select, and as part of this update, the user autocomplete
was changed from a return value of Name to id. This exposed
an existing bug in RT::Users::SimpleSearch when using a return
of id on Pg. Searching for "main.id != ''" generates an error:

    ERROR:  invalid input syntax for type integer

Skip the NOT NULL check for id for all DBs since the id can
never be null.

Fixes: I#38036