Refactor GetCalendarTickets to return a single data structure
The updated code processes the tickets for the current month
and records enough metadata to indicate individual ticket events
and multi-day spans in a single data structure.

This simplifies the code, should make it faster, and creates
a data structure with more metadata, which could be useful if
we switch to a javascript calandar in the future.