Route sorting links back to Open.html page in SelfService
Open.html was created when index.html was updated in 47c40241e3 to
support displaying a dashboard instead of open tickets. When the code
was moved, the BaseURL still pointed to index.html, so the sorting links
didn't work correctly.

Update BaseURL to stay on Open.html when processing sort links.